builds/unix/configure.raw: fix data type sizes
The configure check for sizeof(long) runs a program that ends up deciding that the
value of that parameter is 8, even though it is actually 4 on Emscripten. By
fooling configure into thinking that we're in cross-compilation mode, it uses an
alternate approach that detects the correct value.

This particular solution seems like the right thing to do to prevent other
analogous problems that could potentially pop up.